+++ src/main/java/com/projetb3/api/controller/OrderController.java
@@ -2,6 +2,7 @@ package com.projetb3.api.controller;
 
 import com.projetb3.api.model.Item;
 import com.projetb3.api.model.Order;
+import com.projetb3.api.security.AuthenticationWithJWT;
 import com.projetb3.api.service.ItemService;
 import com.projetb3.api.service.OrderService;
 import org.springframework.http.ResponseEntity;
@@ -10,30 +11,42 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.*;
 
+import static com.projetb3.api.security.AuthenticationWithJWT.verifyJwt;
+import static com.projetb3.api.security.AuthenticationWithJWT.verifySenderOfRequest;
+
 @Controller
 @RequestMapping("/commandes")
 public class OrderController {
 
     private final OrderService orderService;
 
-    public OrderController(OrderService orderService, ItemService itemService) {
+    public OrderController(OrderService orderService) {
         this.orderService = orderService;
     }
 
     @GetMapping
-    public ResponseEntity<Iterable<Order>> getAll() {
-        Iterable<Order> ordersList = orderService.getAll();
-        return ResponseEntity.ok(ordersList);
+    public ResponseEntity<Iterable<Order>> getAll(@RequestHeader("Authentication") final String token) {
+        if (verifySenderOfRequest(token, Optional.empty())) {
+            Iterable<Order> ordersList = orderService.getAll();
+            return ResponseEntity.ok(ordersList);
+        }
+        return ResponseEntity.badRequest().build();
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Order> get(@PathVariable("id") final int id) {
+    public ResponseEntity<Order> get(@PathVariable("id") final int id, @RequestHeader("Authentication") final String token) {
         Optional<Order> order = orderService.get(id);
-        return order.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
+        if(order.isPresent() && verifySenderOfRequest(token, Optional.of(order.get().getUser().getFirstname()))){
+            return order.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
+        }
+        return ResponseEntity.badRequest().build();
     }
 
+    /** !!! faire get les commandes d'un user !!! */
+
     @PostMapping
-    public ResponseEntity<String> processToCreation(@RequestBody Order order) {
+    public ResponseEntity<String> processToCreation(@RequestBody Order order, @RequestHeader("Authentication") final String token) {
+        verifyJwt(token);
         order.fillFields();
         return create(order);
     }
@@ -49,7 +62,7 @@ public class OrderController {
 
     private void decrementItemStock(Order order) {
         int totalPrice = 0;
-        for(Item item : order.getItems()){
+        for (Item item : order.getItems()) {
             orderService.decrementItemStock(item.getId());
             totalPrice += orderService.getPriceOfItem(item.getId());
         }
@@ -62,9 +75,9 @@ public class OrderController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> delete(@PathVariable("id") final int id) {
+    public ResponseEntity<String> delete(@PathVariable("id") final int id, @RequestHeader("Authentication") final String token) {
         Optional<Order> optOrder = orderService.get(id);
-        if (optOrder.isPresent()) {
+        if (optOrder.isPresent() && verifySenderOfRequest(token, Optional.empty())) {
             orderService.delete(id);
             return ResponseEntity.ok().body("La commande a été supprimée.");
         }
@@ -72,9 +85,11 @@ public class OrderController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<String> update(@PathVariable("id") final int id, @RequestBody Order modified) {
+    public ResponseEntity<String> update(@PathVariable("id") final int id,
+                                         @RequestBody Order modified,
+                                         @RequestHeader("Authentication") final String token) {
         Optional<Order> optOrder = orderService.get(id);
-        if (optOrder.isPresent()) {
+        if (optOrder.isPresent() && verifySenderOfRequest(token, Optional.empty())) {
             Order current = optOrder.get();
             if (modified.getNumber() != null) {
                 current.setNumber(modified.getNumber());
